#VMS-2279 refactored 'disconnect from cloud': cloud users removed, 'admin' user enabled first of all
#VMS-3473 fixed: incorrect async call never returned
(VMS-3437) - light client can't view video
(VMS-3449) - Some cameras have invalid colors
(VMS-3452) - lite client: Change quality doesnt work for live mode
=VMS-3111: Minilauncher for MacOS
Access Rights: refactor to move out indirect access helpers
Access Rights: refactor to move out user role calculation
Audit trail layout fix.
BUILD-7 - win x86
Backed out changeset 34b8df918be9
Backed out changeset 6e1ed3310ea5
Cleanup bunch of MSVC compiler warnings
Close xvms_2280_unlink_system branch
Coding style fixes & refactoring.
Custom item view side indents fix.
Default disabled item opacity (30%) is now known as `style::Metrics::kDefaultItemOpacity`
Disable resource tree debug
Disabled Videowall icons
Editable combobox fucus fix
Expert tab was limiting Camera Settings width due to a long warning label
Fix missing nullptr_t with clang++ and libstdc++.
Fix paxton installer build
Fixed falling out of Live mode for MJPEG.
Highlighting checked resources in the tree (used in dialogs)
Internal: Compiling fixed for Linux.
Internal: Fixed deps in mobile_client MSVC Solution.
Internal: Fixed mobile_client args processing.
Internal: Temporary fix for mobile_client crash on Windows; todo added.
Internal: mobile_client: flag_config option added to force non_lite_mode.
Linix compile fix
MOBILE-231: Allow lite client layout creation.
MOBILE-231: Compilation fix
MOBILE-231: Fix camera thumbnails on the lite client controller screen.
MOBILE-231: Fix initial lite client controller state.
MOBILE-231: Fix invalid layout direction.
MOBILE-231: Fix startLiteClient method invokation.
MOBILE-231: Fix switch checkstate bug.
MOBILE-231: Implement starting/stopping states.
MOBILE-231: Inmplement lite client server offline state.
MOBILE-231: Make lite client layout editing work.
MOBILE-231: Move QnLiteClientHandler initialiation to main.cpp
MOBILE-231: Move lite client layout helper functions to a separate class.
MOBILE-231: New lite client resources screen implementation.
MOBILE-231: Open the proper layout in the lite client at startup.
MOBILE-231: Remove debug output.
MOBILE-231: Rewrite lite client controller state to enum instead of bool flag.
MOBILE-231: Simplify mode selector.
MOBILE-231: Update lite client offline icon
MOBILE-266: Fix items aspect ratio in lite client.
MOBILE: A huge simplification of video scaling/rotating.
MOBILE: Android fix
MOBILE: Android fix.
MOBILE: Another fix for json transactions serialization.
MOBILE: Fix Android title bar padding.
MOBILE: Fix camera status text alignment.
MOBILE: Fix cloud connect errors handling.
MOBILE: Fix openssl version for Android.
MOBILE: Focus improvements.
MOBILE: Force focus to the current page if it is lost.
MOBILE: Further focus control improvements.
MOBILE: Make it possible to completely hide tool bar from page.
MOBILE: Make switch disabled state more noticeable.
MOBILE: Remove DISABLE_FFMPEG macro.
MOBILE: Remove ENABLE_ARCHIVE macro.
MOBILE: Remove unnecessary assert.
MOBILE: Restart player if cameraId is changed.
MOBILE: Save client core settings when recent connections is changed.
MOBILE: Speed up cameras list scrolling.
MOBILE: Update camera offline icons.
MS-3439: Fisheye settings redesign
Minor fix.
Minor reformatting & tiny refactoring.
Minor visual fix.
New icons.
QnMessageBox has fixed size by default now
QnNxStyle::standardPixmap --> QnNxStyle::standardIcon (due to Qt update)
QnResourcePreviewWidget with name and description labels grouped into QnResourceDetailsWidget
Resource selection dialog visual improvements.
Resource tree (in dialogs) visual layout fix
Resource tree: Fix current user display
Review fix
Shared layouts icons
Some fixes.
Style fix
Tiny UI fix.
Todo added
Tree View: Hide 'Layouts' node for current user if it is empty
User list drawing refactoring and keyboard interaction fix.
VMS-2266: Camera states
VMS-2266: Camera states. Build fix
VMS-2272 - Unlink system from the cloud behavior - server side
VMS-2280 - Unlink system from the cloud behavior - client side
VMS-3006 - Only owner can link/unlink system to the Cloud
VMS-3111: Minilauncher for MacOS
VMS-3117 - PTZ Presets must be managed by Edit Camera Settings permission (Client Side)
VMS-3173 - Screenshots at the 'Select Devices...' window mustn't lead to resizing of the device list
VMS-3237 - Make IO module require User Input permission (Client Side)
VMS-3265 - Modify local layout for roled user
VMS-3268 - Delete shared layout under roled user
VMS-3276 - Display user role as an additional information for user in the tree, not name
VMS-3281: indirect access providers in accessible resources list
VMS-3317 - Add 'Reset all warnings' button
VMS-3336 - Add 'Manage Account' link to the 'User Settings' window for logged in cloud user.
VMS-3337 - Cloud 'Manage Account...' link doesn't work
VMS-3348 - Tree view - custom user
VMS-3350 - Tree view - custom role
VMS-3428 - Admin user has no permission to remove storage
VMS-3439: Fisheye settings redesign
VMS-3440: Fix crashes and hangs on Mac os
VMS-3470 - Check cloud owner password in the 'Unlink' dialog
VMS-3472:Add applauncher for MacOS installer and zip-updates
VMS: Compilation fix.
VMS: Do not save empty-id records to db.
tr fix

<==== Mercurial data ====>
Branch: prod_3.0.0

+++ client/libclient/src/camera/sync_dialog_display.cpp
@@ -4,7 +4,7 @@
 #include "nx/streaming/archive_stream_reader.h"
 #include "plugins/resource/avi/avi_archive_delegate.h"
 
-QnSignDialogDisplay::QnSignDialogDisplay(QnMediaResourcePtr resource): 
+QnSignDialogDisplay::QnSignDialogDisplay(QnMediaResourcePtr resource):
     QnCamDisplay(resource, 0),
     m_mdctx(EXPORT_SIGN_METHOD)
 {
@@ -31,12 +31,12 @@ void QnSignDialogDisplay::finilizeSign()
     QnSignHelper signHelper;
     signFromPicture = signHelper.getSign(lastFrame.data(), calculatedSign.size());
 #else
-    if (m_reader) 
+    if (m_reader)
     {
         QnAviArchiveDelegate* aviFile = dynamic_cast<QnAviArchiveDelegate*> (m_reader->getArchiveDelegate());
         if (aviFile) {
             const char* signPattern = aviFile->getTagValue(QnAviArchiveDelegate::SignatureTag);
-            if (signPattern) 
+            if (signPattern)
             {
                 QByteArray baPattern = QByteArray(signPattern).trimmed();
                 QByteArray magic = QnSignHelper::getSignMagic();
@@ -69,11 +69,11 @@ bool QnSignDialogDisplay::processData(const QnAbstractDataPacketPtr& data)
         m_reader = reader;
 
     QnEmptyMediaDataPtr eofPacket = std::dynamic_pointer_cast<QnEmptyMediaData>(data);
-    
+
     QnAbstractMediaDataPtr media = std::dynamic_pointer_cast<QnAbstractMediaData>(data);
     QnCompressedVideoDataPtr video = std::dynamic_pointer_cast<QnCompressedVideoData>(data);
     QnCompressedAudioDataPtr audio = std::dynamic_pointer_cast<QnCompressedAudioData>(data);
-    if (m_prevSpeed != m_speed) 
+    if (m_prevSpeed != m_speed)
     {
         processNewSpeed(m_speed);
         m_prevSpeed = m_speed;
@@ -102,7 +102,7 @@ bool QnSignDialogDisplay::processData(const QnAbstractDataPacketPtr& data)
         // update digest from current frame
         if (media && media->dataSize() > 4) {
             const quint8* data = (const quint8*) media->data();
-            QnSignHelper::updateDigest(media->context, m_mdctx, data, media->dataSize());
+            QnSignHelper::updateDigest(media->context, m_mdctx, data, static_cast<int>(media->dataSize()));
         }
 #else
         // update digest from previous frames because of last frame it is sign frame itself
@@ -111,7 +111,7 @@ bool QnSignDialogDisplay::processData(const QnAbstractDataPacketPtr& data)
             QnSignHelper::updateDigest(m_prevFrame->context, m_mdctx, data, m_prevFrame->data.size());
         }
 #endif
-        if (video) 
+        if (video)
         {
             video->channelNumber = 0; // ignore layout info
             if (!m_firstFrameDisplayed || ((video->flags & AV_PKT_FLAG_KEY) && qnSyncTime->currentMSecsSinceEpoch() - m_lastDisplayTime > 100)) // max display rate 10 fps
@@ -150,7 +150,7 @@ bool QnSignDialogDisplay::processData(const QnAbstractDataPacketPtr& data)
     return true;
 }
 
-bool QnSignDialogDisplay::canAcceptData() const 
+bool QnSignDialogDisplay::canAcceptData() const
 {
     return QnAbstractDataConsumer::canAcceptData();
 }
